Fix bug

+++ db/migrate/20230826140654_create_emergency_contact.rb
@@ -3,7 +3,7 @@ class CreateEmergencyContact < ActiveRecord::Migration[6.1]
     create_table :emergency_contacts do |t|
       t.string :name
       t.string :relationship
-      t.integer :number
+      t.bigint :number
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
